feat: updated highscore view
